+++ src/api/public/me/me.controller.ts
@@ -26,7 +26,7 @@ import {
 } from '@nestjs/swagger';
 import { Request } from 'express';
 
-import { TokenAuthGuard } from '../../../auth/token-auth.guard';
+import { TokenAuthGuard } from '../../../auth/token.strategy';
 import { NotInDBError } from '../../../errors/errors';
 import { HistoryEntryUpdateDto } from '../../../history/history-entry-update.dto';
 import { HistoryEntryDto } from '../../../history/history-entry.dto';
