+++ src/api/public/notes/notes.controller.ts
@@ -32,7 +32,7 @@ import {
 } from '@nestjs/swagger';
 import { Request } from 'express';
 
-import { TokenAuthGuard } from '../../../auth/token-auth.guard';
+import { TokenAuthGuard } from '../../../auth/token.strategy';
 import {
   AlreadyInDBError,
   ForbiddenIdError,
